refactor: edit comment dialog 분리
