+++ src/components/dialog/EditCommentDialog.tsx
@@ -1,0 +1,51 @@
+import { useState } from "react"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "../../shared/ui"
+import { useDialogStore } from "../../store/dialog"
+import { useSelectedComment } from "../../store/selectedComment"
+
+const EDIT_COMMENT_DIALOG = "EDIT_COMMENT_DIALOG"
+
+export const EditCommentDialog = () => {
+  const { isDialogOpen, setDialogOpen } = useDialogStore()
+  const { selectedComment, setSelectedComment } = useSelectedComment()
+
+  if (!selectedComment) return null
+
+  // 댓글 업데이트
+  const updateComment = async () => {
+    if (!selectedComment) return
+    try {
+      const response = await fetch(`/api/comments/${selectedComment.id}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ body: selectedComment.body }),
+      })
+      const data = await response.json()
+      //   setComments((prev) => ({
+      //     ...prev,
+      //     [data.postId]: prev[data.postId].map((comment) => (comment.id === data.id ? data : comment)),
+      //   }))
+      setDialogOpen(EDIT_COMMENT_DIALOG, false)
+    } catch (error) {
+      console.error("댓글 업데이트 오류:", error)
+    }
+  }
+
+  return (
+    <Dialog open={isDialogOpen(EDIT_COMMENT_DIALOG)} onOpenChange={(open) => setDialogOpen(EDIT_COMMENT_DIALOG, open)}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>댓글 수정</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={selectedComment?.body || ""}
+            onChange={(e) => setSelectedComment({ ...selectedComment, body: e.target.value })}
+          />
+          <Button onClick={updateComment}>댓글 업데이트</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
